#18[refactor] profile 관련 Dto 생성및 UserEntity 매서드 수정
