+++ src/main/java/com/petfam/petfam/dto/user/ProfileUpdateDto.java
@@ -1,0 +1,11 @@
+package com.petfam.petfam.dto.user;
+
+import lombok.Getter;
+
+@Getter
+public class ProfileUpdateDto {
+  private String nickname;
+  private String introduction;
+  private String image;
+
+}
