+++ src/main/java/com/petfam/petfam/entity/User.java
@@ -1,5 +1,6 @@
 package com.petfam.petfam.entity;
 
+import com.petfam.petfam.dto.user.ProfileUpdateDto;
 import com.petfam.petfam.entity.enums.UserRoleEnum;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
@@ -100,5 +101,10 @@ public class User {
   // 닉네임 업데이트
   public void updateNickname(String nickname) {
     this.nickname = nickname;
+  }
+
+  // 프로필 업데이트
+  public void updateProfile(ProfileUpdateDto profileUpdateDto) {
+
   }
 }
