+++ src/main/java/com/petfam/petfam/service/user/UserService.java
@@ -1,10 +1,11 @@
 package com.petfam.petfam.service.user;
 
-import com.petfam.petfam.dto.AdminSignupRequestDto;
-import com.petfam.petfam.dto.ProfileResponseDto;
-import com.petfam.petfam.dto.ProfileUpdateDto;
-import com.petfam.petfam.dto.SigninRequestDto;
-import com.petfam.petfam.dto.UserSignupRequestDto;
+
+import com.petfam.petfam.dto.user.AdminSignupRequestDto;
+import com.petfam.petfam.dto.user.ProfileResponseDto;
+import com.petfam.petfam.dto.user.ProfileUpdateDto;
+import com.petfam.petfam.dto.user.SigninRequestDto;
+import com.petfam.petfam.dto.user.UserSignupRequestDto;
 import com.petfam.petfam.entity.User;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
